Finish implementing add of new comment

+++ server/src/main/java/ir/asparsa/hobbytaste/server/database/repository/StoreCommentRepository.java
@@ -22,6 +22,6 @@ public interface StoreCommentRepository extends JpaRepository<CommentModel, Long
 
     Optional<CommentModel> findById(Long id);
 
-    Optional<CommentModel> findByHashCodeAndStore(int hashCode, StoreModel store);
+    Optional<CommentModel> findByHashCodeAndStore(long hashCode, StoreModel store);
 
 }
